+++ mod/src/main/java/basemod/abstracts/CustomCard.java
@@ -444,10 +444,13 @@ public abstract class CustomCard extends AbstractCard {
 			return;
 		}
 
-		BitmapFont savedFont = FontHelper.cardTitleFont_small;
-		FontHelper.cardTitleFont_small = titleFont;
+		BitmapFont savedFont = FontHelper.cardTitleFont;
+		FontHelper.cardTitleFont = titleFont;
+		Boolean useSmallTitleFont = ReflectionHacks.getPrivate(this, AbstractCard.class, "useSmallTitleFont");
+		ReflectionHacks.setPrivate(this, AbstractCard.class, "useSmallTitleFont", false);
 		SpireSuper.call(sb);
-		FontHelper.cardTitleFont_small = savedFont;
+		ReflectionHacks.setPrivate(this, AbstractCard.class, "useSmallTitleFont", useSmallTitleFont);
+		FontHelper.cardTitleFont = savedFont;
 	}
 
 	private static BitmapFont generateTitleFont(float size) {