+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/screens/DeathScreen/NotFoundFix.java
@@ -1,15 +1,22 @@
 package basemod.patches.com.megacrit.cardcrawl.screens.DeathScreen;
 
+import basemod.ReflectionHacks;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
-import com.megacrit.cardcrawl.screens.DeathScreen;
+import com.megacrit.cardcrawl.screens.GameOverScreen;
+import com.megacrit.cardcrawl.unlock.AbstractUnlock;
 
-@SpirePatch(cls="com.megacrit.cardcrawl.screens.DeathScreen", method="calculateUnlockProgress")
+import java.util.ArrayList;
+
+@SpirePatch(
+		clz=GameOverScreen.class,
+		method="calculateUnlockProgress"
+)
 public class NotFoundFix {
-	public static void Postfix(Object __obj_instance) {
-		DeathScreen screen = (DeathScreen) __obj_instance;
-		if (screen.unlockBundle != null && screen.unlockBundle.size() <= 0) {
+	public static void Postfix(GameOverScreen __instance) {
+		ArrayList<AbstractUnlock> unlockBundle = ReflectionHacks.getPrivate(__instance, GameOverScreen.class, "unlockBundle");
+		if (unlockBundle != null && unlockBundle.size() <= 0) {
 			// game checks for null, not for wrong size
-			screen.unlockBundle = null;
+			ReflectionHacks.setPrivate(__instance, GameOverScreen.class, "unlockBundle", null);
 		}
 	}
 }